Enum mt_main_io::MTokenAction

pub enum MTokenAction {
    Message {
        transaction_id: u64,
        payload: LogicAction,
    },
    UpdateLogicContract {
        mt_logic_code_hash: H256,
        storage_code_hash: H256,
    },
    GetBalance {
        token_id: TokenId,
        account: ActorId,
    },
    GetApproval {
        account: ActorId,
        approval_target: ActorId,
    },
    Clear(H256),
    MigrateStorageAddresses,
}

Sends the contract info about what it should do.

Variants

Message

Fields

transaction_id: u64

Operation transaction id(each new abstract-transaction must increase).

payload: LogicAction

Encoded high-level LogicAction operation.

Handles high-level token operations.

UpdateLogicContract

Fields

mt_logic_code_hash: H256

Unique hash-identifier of logic contract code.

storage_code_hash: H256

Unique hash-identifier of storage contract code.

Updates unique hash-identifier or multitoken storage and logic contract code.

On success, replies with MTokenEvent::Ok.

GetBalance

Fields

token_id: TokenId

Token ID to get the balance.

account: ActorId

Specifies the account whose balance you want to find out.

Returns account token balance.

On success, replies with MTokenEvent::Balance.

GetApproval

Fields

account: ActorId

An account that provides approve.

approval_target: ActorId

An account that is being verified.

Returns status approval for approval_target from account.

On success, replies with MTokenEvent::Approval.

Clear(H256)

Deletes the stored transaction entity with its status by unique hash.

MigrateStorageAddresses

Unimplemented.